Refactor internals to decouple watch strategies and extract IO
