+++ integrations/webpack-4/tests/integration.test.js
@@ -25,13 +25,14 @@ describe('static build', () => {
   })
 })
 
-describe('watcher', () => {
-  test('classes are generated when the html file changes', async () => {
+describe.each([
+  { TAILWIND_MODE: 'watch' },
+  { TAILWIND_MODE: 'watch', TAILWIND_DISABLE_TOUCH: true },
+])('watcher %p', (env) => {
+  test(`classes are generated when the html file changes`, async () => {
     await writeInputFile('index.html', html`<div class="font-bold"></div>`)
 
-    let runningProcess = $('webpack --mode=development --watch', {
-      env: { TAILWIND_MODE: 'watch', TAILWIND_DISABLE_TOUCH: true },
-    })
+    let runningProcess = $('webpack --mode=development --watch', { env })
 
     await waitForOutputFileCreation('main.css')
 
@@ -80,12 +81,10 @@ describe('watcher', () => {
     return runningProcess.stop()
   })
 
-  test('classes are generated when the tailwind.config.js file changes', async () => {
+  test(`classes are generated when the tailwind.config.js file changes`, async () => {
     await writeInputFile('index.html', html`<div class="font-bold md:font-medium"></div>`)
 
-    let runningProcess = $('webpack --mode=development --watch', {
-      env: { TAILWIND_MODE: 'watch', TAILWIND_DISABLE_TOUCH: true },
-    })
+    let runningProcess = $('webpack --mode=development --watch', { env })
 
     await waitForOutputFileCreation('main.css')
 
@@ -148,12 +147,10 @@ describe('watcher', () => {
     return runningProcess.stop()
   })
 
-  test('classes are generated when the index.css file changes', async () => {
+  test(`classes are generated when the index.css file changes`, async () => {
     await writeInputFile('index.html', html`<div class="font-bold btn"></div>`)
 
-    let runningProcess = $('webpack --mode=development --watch', {
-      env: { TAILWIND_MODE: 'watch', TAILWIND_DISABLE_TOUCH: true },
-    })
+    let runningProcess = $('webpack --mode=development --watch', { env })
 
     await waitForOutputFileCreation('main.css')
