+++ src/jit/processTailwindFeatures.js
@@ -1,0 +1,15 @@
+import expandTailwindAtRules from './lib/expandTailwindAtRules'
+import expandApplyAtRules from './lib/expandApplyAtRules'
+import evaluateTailwindFunctions from '../lib/evaluateTailwindFunctions'
+import substituteScreenAtRules from '../lib/substituteScreenAtRules'
+import collapseAdjacentRules from './lib/collapseAdjacentRules'
+
+export default function processTailwindFeatures(context) {
+  return function (root, result) {
+    expandTailwindAtRules(context)(root, result)
+    expandApplyAtRules(context)(root, result)
+    evaluateTailwindFunctions(context)(root, result)
+    substituteScreenAtRules(context)(root, result)
+    collapseAdjacentRules(context)(root, result)
+  }
+}
