+++ src/lib/substituteScreenAtRules.js
@@ -1,7 +1,7 @@
 import _ from 'lodash'
 import buildMediaQuery from '../util/buildMediaQuery'
 
-export default function ({ theme }) {
+export default function ({ tailwindConfig: { theme } }) {
   return function (css) {
     css.walkAtRules('screen', (atRule) => {
       const screen = atRule.params
