+++ tests/screenAtRule.test.js
@@ -3,7 +3,7 @@ import plugin from '../src/lib/substituteScreenAtRules'
 import config from '../stubs/defaultConfig.stub.js'
 
 function run(input, opts = config) {
-  return postcss([plugin(opts)]).process(input, { from: undefined })
+  return postcss([plugin({ tailwindConfig: opts })]).process(input, { from: undefined })
 }
 
 test('it can generate media queries from configured screen sizes', () => {
